Update critter.hpp
Updated critter header with delete flag to indicate when object should be deleted. Needed because move functions within object can copy data to a new location, but can't delete current location.
Added move(), breed(), and starve() as pure virtual functions. Unsure if starve() is needed, but rubric seems to suggest it is.